reolsved

+++ app/Http/Controllers/TripController.php
@@ -536,7 +536,9 @@ class TripController extends Controller
                         ];
 
                         $result = $this->markOptimumFuelStations($tripDetailResponse);
-
+                        if($result==false){
+                            return response()->json(['status'=>404,'message'=>'no fuel station in range','data'=>(object)[]]);
+                        }
                         foreach ($result as  $value) {
                          
 
@@ -904,6 +906,9 @@ class TripController extends Controller
                 ];
 
                 $result = $this->markOptimumFuelStations($tripDetailResponse);
+                if($result==false){
+                    return response()->json(['status'=>404,'message'=>'no fuel station in range','data'=>(object)[]]);
+                }
                     //$result = $this->findOptimalFuelStation($startLat, $startLng, $truckMpg, $currentFuel, $matchingRecords, $endLat, $endLng);
                     FuelStation::where('trip_id', $trip->id)->delete();
                     foreach ($result as  $value) {
@@ -1121,7 +1126,9 @@ class TripController extends Controller
         // Separate stations into in-range and out-of-range based on truck's fuel capacity
         $fuelStationsInRange = $fuelStations->filter(fn($fs) => $fs['distanceFromStart'] < $truckTravelableDistanceInMiles);
         $fuelStationsOutsideRange = $fuelStations->reject(fn($fs) => $fs['distanceFromStart'] < $truckTravelableDistanceInMiles);
-
+        if($fuelStationsInRange->isEmpty()){
+            return false;
+        }
         // Find the cheapest stations
         $firstCheapestInRange = $fuelStationsInRange->sortBy('price')->first();
